+++ angular/src/app/data/bathroom.model.ts
@@ -1,0 +1,12 @@
+/**
+ * Represents the _Bathroom_ model
+ *
+ * ```yaml
+ * id: string;
+ * fixture: number;
+ * ```
+ */
+export interface Bathroom {
+  id: string;
+  fixture: number;
+}