+++ angular/src/app/data/bedroom.model.ts
@@ -1,0 +1,14 @@
+/**
+ * Represents the _Bedroom_ model
+ *
+ * ```yaml
+ * id: string;
+ * count: number;
+ * type: string;
+ * ```
+ */
+export interface Bedroom {
+  id: string;
+  count: number;
+  type: string;
+}