+++ angular/src/app/data/location.model.ts
@@ -1,0 +1,18 @@
+import { Address } from './address.model';
+
+/**
+ * Represents the _Location_ model
+ *
+ * id: string;
+ * address: Address;
+ * latitude: string;
+ * locale: string;
+ * longitude: string;
+ */
+export interface Location {
+  id: string;
+  address: Address;
+  latitude: string;
+  locale: string;
+  longitude: string;
+}