+++ angular/src/app/data/rental-unit.model.ts
@@ -1,0 +1,23 @@
+import { Bathroom } from './bathroom.model';
+import { Bedroom } from './bedroom.model';
+
+/**
+ * Represents the _RentalUnit_ model
+ *
+ * ```yaml
+ * id: string;
+ * bathrooms: Bathroom[];
+ * bedrooms: Bedroom[];
+ * name: string;
+ * occupancy: number;
+ * type: string;
+ * ```
+ */
+export interface RentalUnit {
+  id: string;
+  bathrooms: Bathroom[];
+  bedrooms: Bedroom[];
+  name: string;
+  occupancy: number;
+  type: string;
+}