+++ angular/src/app/data/rental.model.ts
@@ -1,16 +1,16 @@
+import { RentalUnit } from './rental-unit.model';
+
 /**
  * Represents the _Rental_ model
  *
  * ```yaml
  * id: string;
  * name: string;
- * occupancy: number;
- * type: string;
+ * rentalUnit: RentalUnit;
  * ```
  */
 export interface Rental {
   id: string;
   name: string;
-  occupancy: number;
-  type: string;
+  rentalUnit: RentalUnit;
 }