+++ angular/src/app/data/review.model.ts
@@ -1,0 +1,20 @@
+/**
+ * Represents the _Review_ model
+ *
+ * ```yaml
+ * id: string;
+ * accountId: string;
+ * hotelId: string;
+ * comment: string;
+ * dateCreated: Date;
+ * rating: number;
+ * ```
+ */
+export interface Review {
+  id: string;
+  accountId: string;
+  hotelId: string;
+  comment: string;
+  dateCreated: Date;
+  rating: number;
+}